refactor: move validation to its own layer
issue: #26
