+++ src/presentation/controllers/login/login-controller-protocols.ts
@@ -1,3 +1,2 @@
 export * from '../../protocols'
 export * from '../../../domain/usecases/authentication'
-export * from '../../protocols/validation'
