+++ src/presentation/controllers/signup/signup-controller-protocols.ts
@@ -2,4 +2,3 @@ export * from '../../protocols'
 export * from '../../../domain/models/account'
 export * from '../../../domain/usecases/add-account'
 export * from '../../../domain/usecases/authentication'
-export * from '../../protocols/validation'
