+++ src/validation/validators/compare-fields-validation.ts
@@ -1,5 +1,5 @@
-import { Validation } from '../../protocols/validation'
-import { InvalidParamError } from '../../errors'
+import { Validation } from '../../presentation/protocols'
+import { InvalidParamError } from '../../presentation/errors'
 
 export class CompareFieldsValidation implements Validation {
   constructor (
